pyproject.toml
[project]
name = "pretalx-oidc"
dynamic = ["version"]
description = "pretalx plugin for oidc auth"
readme = "README.rst"
license = {text = "Apache Software License"}
keywords = ["pretalx"]
authors = [
  {name = "Your name", email = "your-email@example.org"},
]
maintainers = [
  {name = "Your name", email = "your-email@example.org"},
]

dependencies = []

[project.urls]
homepage = "https://github.com/snakedev24/pretalx-oidc-plugin"
repository = "https://github.com/snakedev24/pretalx-oidc-plugin.git"

[project.entry-points."pretalx.plugin"]
pretalx_oidc = "pretalx_oidc:PretalxPluginMeta"

[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools"]

[tool.setuptools]
include-package-data = true

[tool.setuptools.dynamic]
version = {attr = "pretalx_oidc.__version__"}

[tool.setuptools.packages.find]
include = ["pretalx*"]

[tool.check-manifest]
ignore = [".*"]

[tool.poetry.dependencies]
pretalx_oidc_auth = "*"

[tool.poetry.dev-dependencies]
mozilla-django-oidc = "*"

[tool.pretalx_oidc_auth]
OIDC_RP_CLIENT_ID = "your-client-id"
OIDC_RP_CLIENT_SECRET = "your-client-secret"
OIDC_RP_CLIENT_SECRET_ENCODED = true
OIDC_OP_JWKS_ENDPOINT = "https://your-oidc-provider.com/jwks"
OIDC_OP_AUTHORIZATION_ENDPOINT = "https://your-oidc-provider.com/authorize"
OIDC_OP_TOKEN_ENDPOINT = "https://your-oidc-provider.com/token"
OIDC_OP_USER_ENDPOINT = "https://your-oidc-provider.com/userinfo"
OIDC_OP_LOGOUT_URL_METHOD = "auth_logout"
